+++ src/views/song/Info/comp/InfoLabel.tsx
@@ -1,0 +1,5 @@
+import type { ParentProps } from "solid-js"
+
+export function InfoLabel(props: ParentProps) {
+	return <span class="text-sm text-tertiary">{props.children}</span>
+}
